auto-attendance-system/ai_service/requirements.txt
# Auto Attendance System - AI Service Dependencies
# Phase 4: Face Recognition & Attendance Automation

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Computer Vision & Deep Learning
opencv-python==4.8.1.78
torch==2.1.0
torchvision==0.16.0
facenet-pytorch==2.5.3

# Image Processing
Pillow==10.0.1
numpy==1.24.4

# Vector Search & Similarity
faiss-cpu==1.7.4
scikit-learn==1.3.2

# Database & HTTP
requests==2.31.0
aiohttp==3.9.0
pymongo==4.6.0

# Utilities & Logging
loguru==0.7.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2